requirements.txt
APScheduler==3.7.0
beautifulsoup4==4.9.1
Flask==1.1.2
line-bot-sdk==1.17.0
lxml==4.5.2
requests==2.24.0
selenium==3.141.0
twstock==1.3.1


